Moving auth to amp user service
